+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/state/extension/WebExtensionPromptRequest.kt
@@ -64,11 +64,13 @@ sealed class WebExtensionPromptRequest {
              * Value type that represents a request for an optional permissions prompt.
              * @property extension The [WebExtension] that requested the dialog to be shown.
              * @property permissions The optional permissions to list in the dialog.
+             * @property origins The optional host permissions to list in the dialog.
              * @property onConfirm A callback indicating whether the permissions were granted or not.
              */
             data class Optional(
                 override val extension: WebExtension,
                 val permissions: List<String>,
+                val origins: List<String>,
                 val onConfirm: (Boolean) -> Unit,
             ) : Permissions(extension)
         }